\name{cumulativeCA}
\alias{cumulativeCA}
\alias{cumulativeca}
\alias{cumulative}
\alias{CA}

\title{
	Cumulative contribution abundance of genes.
}
\description{
	We present a tool to calculate the cumulative contribution abundance of genes in cancer research.
}
\usage{
	cumulativeCA(file = NULL, Filetype = c('MAF','VCF'), Pfile = NULL, Sfile = NULL, sample.id = 'Tumor_Sample_Barcode', chr = 'Chromosome', pos = 'Start_position', ref = 'Reference_Allele', alt = 'Tumor_Seq_Allele2', Hugo = 'Hugo_Symbol', AnalCOSMICSigType = c('SBS','DBS','ID'), genome.build = c("Ch37","Ch38"),  groupFile = NULL, geneListSortFile = NULL, ID.mc.cores = 2, ID.row83 = TRUE, plot = TRUE) 	
}
\arguments{
  \item{file}{       
     A file with non-silent mutations obtained by annotation of mutation locations by Oncotator or other annotation software.
  }
  \item{Filetype}{       
     The format of the input dataset, contains MAF and VCF. File with 'MAF' form was generated by Oncotator, and File with 'VCF' form must have at least six columns with the colnames of "Hugo_Symbol	CHROM	POS	REF	ALT	SAMPLENAME", for SBS  as: 1    1432333       T       C	T1, for DBS  as: 1    1432333       TA       CG	T1, for ID  as: 1    1432333       T       -	T1.
  }
  \item{Pfile}{       
     A mutational signature matrix file P with a format like the output result of RNMF software.
  }
   \item{Sfile}{       
     A abundance fractions matrix S with a format like the output result of RNMF software.
  }
  \item{sample.id}{       
     Column name in the mutation file corresponding to the Sample ID.	
  }
  \item{chr}{       
     Column name in the mutation file corresponding to the chromosome.
  }  
  \item{pos}{       
     Column name in the mutation file corresponding to the mutation position.
  }  
  \item{ref}{       
     Column name in the mutation file corresponding to the reference base.
  }
  \item{alt}{       
     Column name in the mutation file corresponding to the alternate base.
  }
  \item{Hugo}{       
     Column name in the mutation file corresponding to the gene.
  }    
  \item{AnalCOSMICSigType}{       
     Type of mutation based on PCAWG project, canbe SBS, DBS or ID.
  }
  \item{genome.build}{       
     The version of the reference genome used in the mutation set.
  }
  \item{groupFile}{       
     Group files with two columns: 'SampleID', 'Group'.
  }    
  \item{geneListSortFile}{ 
	Sorted gene list with two columns: 'Hugo_Symbol', 'Chromosome', such as "KDSR      chr18".
  }   
  \item{ID.mc.cores}{       
     The number of cores to use in ID type, i.e. at most how many child processes will be run simultaneously. Must be exactly 1 on Windows (which uses the master process).
  }
  \item{ID.row83}{       
    Whether to keep 83 properties for ID type. 
  }
  \item{plot}{       
    Whether to plot the image. 
  }  
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	cumulativeCA(file = 'Inputtest.maf', Pfile = 'SignatureComposition.12.Normalized.txt', Sfile = 'SampleContribution.12.Normalized.txt', sample.id = 'Tumor_Sample_Barcode', chr = 'Chromosome', pos = 'Start_position', ref = 'Reference_Allele', alt = 'Tumor_Seq_Allele2', Hugo = 'Hugo_Symbol', AnalCOSMICSigType = 'SBS')
}
