% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{face}
\alias{face}
\title{An image of a corrupted face image.}
\format{A data matrix containing 192 rows and 168 columns}
\description{
A 192 by 168 data matrix storing pixel intensities of a corrupted face image.
The original image was taken from Yale Face Database B [Georghiades et al. 2001]
and manually corrupted.
}
\keyword{datasets}

