% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnnmf-package.r
\name{rnnmf-NEWS}
\alias{rnnmf-NEWS}
\title{News for package 'rnnmf':}
\description{
News for package \sQuote{rnnmf}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\rnnmf}{\CRANpkg{rnnmf}}
}
\section{\rnnmf{} Version 0.3.1 (2026-07-15) }{

\itemize{
\item CRAN emergency release (vignette cran check fails).
}
}

\section{\rnnmf{} Initial Version 0.3.0 (2024-10-30) }{

\itemize{
\item first CRAN release.
\item changed name from rnmf to rnnmf.
}
}

