% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Nnls}
\alias{Nnls}
\title{Performs Non-negative least square regression on a matrix.}
\usage{
Nnls(M = M, C = C)
}
\arguments{
\item{M}{Design matrix. \code{M}}

\item{C}{Response matrix \code{C}}
}
\value{
output W, a non-negative matrix (To be finished)
}
\description{
Given two matrices M, C >= 0, find W >= 0 such that ||MW - C||_2^2 is minimized.
(This parts to be finished)
}
\examples{
#R code here showing how your function works (To be finished)
}
\keyword{keywords}

