% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnnmf-package.r
\docType{package}
\name{rnnmf}
\alias{rnnmf-package}
\alias{rnnmf}
\title{regularized non-negative matrix factorization}
\description{
Regularized Non-negative Matrix Factorization.
}
\note{
This package provides proof of concept code which is unlikely to be fast
or robust, and may not solve the optimization problem at hand. User assumes
all risk.

This package is maintained as a hobby.
}
\section{Legal Mumbo Jumbo}{


rnnmf is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\references{
Lee, Daniel D. and Seung, H. Sebastian. "Algorithms for Non-negative Matrix 
Factorization." Advances in Neural Information Processing Systems 13 (2001):
556--562.
\url{http://papers.nips.cc/paper/1861-algorithms-for-non-negative-matrix-factorization.pdf}

Pav, S. E. "An Iterative Algorithm for Regularized Non-negative Matrix Factorizations."
Arxiv pre-print (2024).  \doi{10.48550/arXiv.2410.22698}

Pav, Steven E. "System and method for unmixing spectroscopic observations with nonnegative 
matrix factorization." US Patent 8140272, 2012.
\url{https://patents.google.com/patent/US8140272B2/en?oq=8140272}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shabbychef/rnnmf}
  \item Report bugs at \url{https://github.com/shabbychef/rnnmf/issues}
}

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

\strong{Maintainer}: Steven E. Pav \email{shabbychef@gmail.com} (\href{https://orcid.org/0000-0002-4197-6195}{ORCID})

}
\keyword{internal}
\keyword{package}
