\name{SigsInput}
\alias{SigsInput}
\alias{sigsinput}

\title{
	Convert the mutation dataset into input datas for the mutational signature.
}

\description{
	Convert the mutation dataset into input datas for the mutational signature. The format of the input dataset contains MAF and VCF. We used classifications for each type of mutation from PCAWG project. For SBSs, the primary classification comprised 96 classes (available at https://cancer.sanger.ac.uk/cosmic/signatures/SBS) constituted by the 6 base substitutions C>A, C>G, C>T, T>A, T>C and T>G (in which the mutated base is represented by the pyrimidine of the base pair), plus the flanking 5′ and 3′ bases. In some analyses, two flanking bases 5′ and 3′ to the mutated base were considered (producing 1,536 classes) or mutations within transcribed genome regions were selected and classified according to whether the mutated pyrimidine fell on the transcribed or untranscribed strand (producing 192 classes). We also derived a classification for DBSs (78 classes; available at https://cancer.sanger.ac.uk/cosmic/signatures/DBS). Indels were classified as deletions or insertions and—when of a single base—as C or T, and according to the length of the mononucleotide repeat tract in which they occurred. Longer indels were classified as occurring at repeats or with overlapping microhomology at deletion boundaries, and according to the size of indel, repeat and microhomology (83 classes; available at https://cancer.sanger.ac.uk/cosmic/signatures/ID).
}

\usage{
	SigsInput(file = NULL, Filetype = c('MAF','VCF'), AnalCOSMICSigType = c('SBS','DBS','ID'), genome.build = c("Ch37","Ch38"), sample.id = 'Tumor_Sample_Barcode', chr = 'Chromosome', pos = 'Start_position', ref = 'Reference_Allele', alt = 'Tumor_Seq_Allele2', ID.mc.cores = 1, ID.row83 = TRUE)
}

\arguments{
  \item{file}{       
     A mutation file obtained by annotation of mutation locations by Oncotator or other annotation software.
  }
  \item{Filetype}{       
     The format of the input dataset, contains MAF and VCF. File with 'MAF' form was generated by Oncotator, and File with 'VCF' form must have at least five columns with the colnames of "CHROM	POS	REF	ALT	SAMPLENAME", for SBS  as: 1    1432333       T       C	T1, for DBS  as: 1    1432333       TA       CG	T1, for ID  as: 1    1432333       T       -	T1.
  }
  \item{AnalCOSMICSigType}{       
     Type of mutation based on PCAWG project, canbe SBS, DBS or ID.
  }
  \item{genome.build}{       
     The version of the reference genome used in the mutation set.
  }
  \item{sample.id}{       
     Column name in the mutation file corresponding to the Sample ID.	
  }
  \item{chr}{       
     Column name in the mutation file corresponding to the chromosome.
  }  
  \item{pos}{       
     Column name in the mutation file corresponding to the mutation position.
  }  
  \item{ref}{       
     Column name in the mutation file corresponding to the reference base.
  }
  \item{alt}{       
     Column name in the mutation file corresponding to the alternate base.
  }
  \item{ID.mc.cores}{       
     The number of cores to use in ID type, i.e. at most how many child processes will be run simultaneously. Must be exactly 1 on Windows (which uses the master process).
  }
  \item{ID.row83}{       
    Whether to keep 83 properties for ID type. 
  }
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	SigsInput(file = 'Oncotator.HK100.maf', Filetype = 'MAF', AnalCOSMICSigType = 'SBS', genome.build = "Ch37", sample.id = 'Tumor_Sample_Barcode', chr = 'Chromosome', pos = 'Start_position', ref = 'Reference_Allele', alt = 'Tumor_Seq_Allele2')
}