% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbols.R
\docType{data}
\name{Symbols}
\alias{Symbols}
\title{30 images containing 4 different symbols.}
\format{A data matrix containing 30 columns and 5625 rows.}
\description{
A 5625 by 30 matrix. Each column stores vectorized pixels of a 75 by 75 image.
Every image contains one or more of four shapes: circle, square, triangle and cross.
The pixel intensities are standarized to 0 to 1.
}
\keyword{datasets}

