% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbols_c.R
\docType{data}
\name{Symbols_c}
\alias{Symbols_c}
\title{30 Images Containing 4 Different Symbols with Corruptions.}
\format{A 5625 by 30 numeric matrix.}
\description{
A 5625 by 30 matrix. Each column stores vectorized pixels of a 75 by 75 image.
Every image contains one or more of four shapes: circle, square, triangle and cross.
The pixel intensities are standarized to 0 to 1. Box corruptions are randomly added.
Run see(Symbols_c) to visualize.
}
\keyword{datasets}

