% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tumor.R
\docType{data}
\name{tumor}
\alias{tumor}
\title{An Image of 3 Simulated Tumors.}
\format{A 70 by 70 numeric matrix.}
\description{
A 70 by 70 data matrix storing pixel intensities of three tumors that are normalized to 0 to 1.
}
\keyword{datasets}

