% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Tumor.R
\docType{data}
\name{Tumor}
\alias{Tumor}
\title{An image of 3 simulated tumors.}
\format{A data matrix containing 70 columns and 70 rows.}
\description{
A 70 by 70 data matrix storing pixel intensities that are normalized to 0 to 1.
}
\keyword{datasets}

